+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 ### Added
+- config_file, get_config()
+- bird_cmd
 - peer fields: description, address, asn
 - route fields: interface, source, time